同步dev

+++ tool/update-upgrade/ss-do-upgrade.sh
@@ -1,7 +1,12 @@
 #!/bin/bash
+if [ "$(id -u)" != "0" ] ; then
+	pkexec "$0" "$@"
+	exit
+fi
 trap "rm -f  /tmp/spark-store/upgradeStatus.txt" EXIT
 source /opt/durapps/spark-store/bin/bashimport/transhell.amber
 load_transhell_debug
+
 
 function get_name_from_desktop_file() {
 	local app_name_in_desktop
